// src/constants/enums/target-group.constants.ts
export enum TargetGroup {
  BmiMore25 = 'bmiMore25',
  DiabetesType2 = 'diabetesType2',
  DiabetesType1 = 'diabetesType1',
  PreDiabetes = 'preDiabetes',
  InsulinResistance = 'insulinResistance',
  PolycysticOvarianSyndrome = 'polycysticOvarianSyndrome',
  HighBloodPressure = 'highBloodPressure',
  HeartDisease = 'heartDisease',
  HighCholesterolTriglycerides = 'highCholesterolTriglycerides',
  OverweightOrObese = 'overweightOrObese',
  SleepApnea = 'sleepApnea',
  AcidRefluxHeartburn = 'acidRefluxHeartburn',
  FattyLiverDisease = 'fattyLiverDisease',
  ArthritisJointPain = 'arthritisJointPain',
  None = 'none',
  All = 'all',
  AnxietyDepression = 'anxietyDepression',
  FemalesMore45 = 'femalesMore45',
  FemalesBelow45 = 'femalesBelow45',
  Females = 'females',
  MalesMore45 = 'malesMore45',
  MalesBelow45 = 'malesBelow45',
  Males = 'males',
}